refactor: move builtin command handlers to dedicated module
Extract handle_new, handle_edit, and handle_list from main.rs into
src/builtin.rs for better code organization.
